[fix/89-getChallengeDetail] 챌린지 상세 조회 로직 구현

+++ src/main/java/ollie/wecare/challenge/entity/Challenge.java
@@ -45,9 +45,7 @@ public class Challenge extends BaseEntity {
     @OneToMany(fetch = FetchType.LAZY)
     private List<User> participants;
 
-    private Long totalNum;
+    private Integer totalNum;
 
-    private Long attendanceRate;
-
-
+    private Integer attendanceRate;
 }
